Update PlayFabPluginManager.cpp
When doing std::move explicitly to add ownership, the emptied value is only the local version of the shared ptr. These moves are OKAY, but they could be removed entirely if we made the shared_ptr references instead, which DOESN'T increment a strong ref count until we copy the object to our own local version of an object. So this change if it needs to happen must coincide with a potential public change :(