Made changes based on comments
Deb wanted a yield instruction to guarantee the threads we are testing will exit the control block of execution, requiring context switching per thread that gets made.

Made most of the statics local variables with the exception of the threadpool, which needs to wait for the final callback before it can join all the threads and then clear itself (this has to be on test teardown).