Make FromJsonUtilT() predictable for time_t.
mktime() and _mkgmtime() assume tm's tm_isdst is set to a deliberate,
known value.  When it's undefined-- and std::get_time() doesn't write to
it-- the result is unpredictable.  A zero value means "DST is not in
effect", positive means "DST is in effect", and a negative value means
"check the current locale/environment for whether or not DST is in
effect".  Adjustments _may_ be made to the resulting value depending on
this behavior.

Fixes https://github.com/PlayFab/XPlatCppSdk/issues/34

Change *should* be a safe minimal that just makes this function's result
predictable/consistent.  But don't consider this change to be
well-tested just yet.

Ideally, this function wouldn't ignore the incoming value's timezone, if
given.