Testing EventsApi
PrivateMember function test potentially Times out due to utilizing a local API instead of a member API (so it may get deleted and the callback can no longer be retrieved).

Now addressing EventsApi. I think in a similar problem we are calling a STATIC api but are using member callbacks. Logically we should use static callbacks to ensure we aren't doing anything with member variables in this particular test.