Adding threading tests for the event pipeline (#44)
* Adding threading tests for the event pipeline

We are considering a refactor of the PlayFabEventPipeline which is expected to be threadsafe. We don't have tests that cover this scenario, so I added 3 tests: 1 basic 5 threads with 6 events per thread, and then a 30 thread with 2 events per and test with 2 threads and 15 events per.

* Update PlayFabEventTest.cpp

commenting out something that can produce warnings

* Made changes based on comments

Deb wanted a yield instruction to guarantee the threads we are testing will exit the control block of execution, requiring context switching per thread that gets made.

Made most of the statics local variables with the exception of the threadpool, which needs to wait for the final callback before it can join all the threads and then clear itself (this has to be on test teardown).

* Update PlayFabEventTest.cpp

missed removing an excess size check

* Addresses some comments

make the counter atomic and decrease to 0 instead of check for a second counter. Make the threadPool a non-static member of the PlayFabEventTest instead of a static member.

* Update PlayFabEventTest.cpp

adding some spacing for lambda readability.

* Update PlayFabEventTest.h

missed this in a previous fix

* Update PlayFabEventTest.h

quick ordering fix

* Update PlayFabEventTest.cpp

Jenkins failed these two tests. Why?

* Update PlayFabEventTest.cpp

add one problematic test back.

Expected it will time out like last time? Why would using a non static callback make this test time out as opposed to its lambda or static test counter parts?

* Update PlayFabEventTest.cpp

attempting to move eventAPI to be a member rather than a local variable (maybe this is going out of scope?)

* Testing EventsApi

PrivateMember function test potentially Times out due to utilizing a local API instead of a member API (so it may get deleted and the callback can no longer be retrieved).

Now addressing EventsApi. I think in a similar problem we are calling a STATIC api but are using member callbacks. Logically we should use static callbacks to ensure we aren't doing anything with member variables in this particular test.

* Update PlayFabEventTest.cpp

static functions still don't end the test. Attempting to utilize lambdas for this case?

* Update PlayFabEventTest.cpp

cut number of events down.

* Update PlayFabEventTest.cpp

increasing number of events (to two).

* Update PlayFabEventTest.cpp

Weird to see that 2 events makes this not work? But 1 event will?

This test actually emits 4 total events. But apparently Jenkins can't pump out more than 1 at a time?

* Update PlayFabEventTest.cpp

missed a var when commenting code out.

* Final cleanup

removed lambdas, we will push just 1 event now.

* Update PlayFabEventTest.cpp

bad merge