Marks most un-changed shared_ptr parameters to const refs (#38)
* Marks most un-changed shared_ptr parameters to const refs

Most of these were being passed by value, therefore, copying the entire object whenever it was called. Passing the shared_ptr by const ref is still an unchanged standard. I'm leaving certain refs to shared_ptrs since we are writing to that object in the function, and we aren't passing by value, so we shouldn't suffer any copy pentalties.

* Reverts some code

There were some disagreements in favor of not breaking public interfaces that may require the refactor we wanted to avoid.

* Update PlayFabEvent.h

This is the public funciton callback that should probably not be changed.