Update PlayFabAndroidHttpPlugin.cpp (#63)
* Update PlayFabAndroidHttpPlugin.cpp

We never added a requestId to the requestTask on Android to begin with.

* Adding requestId as a parameter to ProcessResponse

This matches our Windows signature, so we should be reading the requestId before this function gets called. adding a temp requestId while I pull out the correct id from the request task (need to potentially parse this  out with the jniEnv)

* Add header get parsing java functions

We need the ability to parse the Headers and explicitly look for the X-RequestId through the jni.

* Update PlayFabAndroidHttpPlugin.cpp

This part is no longer needed as we are adding it before the function directly as a parameter to ProcessResponse (and parsed before passing it in).

* Update PlayFabAndroidHttpPlugin.cpp

removing excess whitespace

* Removing requestId as a parameter

There's a SetRequestId function for the container that we can utilize instead and directly add it when parsed instead of passing it around.

* Update PlayFabAndroidHttpPlugin.cpp

missed a semi colon

* Update PlayFabAndroidHttpPlugin.cpp

remove unneeded requestId

* Update PlayFabCallRequestContainer.cpp

Add the error wrapper to know the requestId when it is set.