Update PlayFabApiTest.cpp
This test fails XPlat Android. The test is meant to exercise our time logic, but we end up relying on date time hardware configurations which have proven unreliable between Android (differnt hardware manufacturers like Samsung may not sync the day when time is synced) and PS4 (setting the timezone to London was still off by hours). This test failure is NOT indicating our timing logic is incorrect, but that our time doesn't match the servers.