Update PlayFabQoSApi.cpp
Current implementation can potentially fail in the event of a time out.

A user reported:

"If the future_status status = asyncPingResults[i].wait_for(pingWaitTime); just timeout, the result is never pulled... when the PlayFabQoSApi::GetResult method returns, the vector<future<PingResult>> asyncPingResults variable destruction cause the code to wait for all future tasks not completed yet."

So replacing the initial if with a while in this particular case should be fine.  There is a question though as to whether or not this starts blocking a thread if we are looping in this manner.