adding a lock to pf event pipeline
There has been some internal complaints about inheriting the event pipeline and messing with one of the protected members that we originally only meant for it to be read synchronously. BUT since it is written to on either a worker thread or on a callback (for cleaning up) it requires us to lock the variable whenever it is used (either being read from or written to).