Add additional error checking and helper functions
It's a bit hacky but we have a shipped protected function that I'd rather not change the function signature of (increases size of the change across 1 to 2 more files, maybe every other HttpPlugin file if we change this signature). SO we have an atomic flag to help see if something went wrong with adding default headers.

We seem to rely on using a raw reference to add error info to a unique_ptr. We will first set that raw reference first before quickly terminating and notify user defined callback functions before attempting to fail sending an invalid header.