Presenting two easy options
Here are two easy and non-breaking changes that we can possibly make for our threaded callbacks issue. The 3rd possibility is a bit complex and will need a bit more time to test. Essentially we will need to remove a generated call that validates the container until later in the http process.

WE MAY want to check for threaded callbacks and make this previous behaviour still work, otherwise, we need to honor calling the user's Update callback (pending Responses will have that error).

It will take longer, but this is most likely what they would expect.