Removing A and C, promoting B
Removing code that is no longer necessary in the HandleInvalidSettings function call. Throwing an exception in this case is better to get up front, in a dev's face and say "This will break if you continue down this path, we will end it here, you must fix this particular setting before calling this again"  And they could potentially try/catch their calls to check for the title Id/reset it, to not have to stop and restart the program if they want.