Toddbell/threaded callbacks fix (#99)
Changing empty title id Settings check to throw an exception instead of hand an error back to the users. The error used to return synchronously, which can be problematic with our threadedCallbacks flag (if it did obey this flag, a user may have to wait for a failed call, this change shortcuts this process).